Fix access violation caused by mismatched integer types (long vs size_t) in pyi_[create,remove]_temp_path

+++ bootloader/src/pyi_utils.c
@@ -269,7 +269,7 @@ int pyi_create_temp_path(ARCHIVE_STATUS *status)
 
 // TODO merge unix/win versions of remove_one() and pyi_remove_temp_path()
 #ifdef _WIN32
-static void remove_one(char *fnm, int pos, struct _finddata_t finfo)
+static void remove_one(char *fnm, size_t pos, struct _finddata_t finfo)
 {
 	if ( strcmp(finfo.name, ".")==0  || strcmp(finfo.name, "..") == 0 )
 		return;
@@ -291,8 +291,8 @@ void pyi_remove_temp_path(const char *dir)
 {
 	char fnm[PATH_MAX+1];
 	struct _finddata_t finfo;
-	long h;
-	int dirnmlen;
+	intptr_t h;
+	size_t dirnmlen;
 	strcpy(fnm, dir);
 	dirnmlen = strlen(fnm);
 	if ( fnm[dirnmlen-1] != '/' && fnm[dirnmlen-1] != '\\' ) {
